docs(error): remove stale FormatAPIError/error_helper.go references (#300)
error_helper.go and FormatAPIError() were deleted when the provider
migrated to CheckResponseErr()/ProviderError. This commit removes all
remaining references from the four doc files that still mentioned the
legacy code, and consolidates the orphaned billingPeriodFromAPI test
from error_helper_test.go into tag_helpers_test.go where it belongs.

Closes #300
